backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
asyncio==4.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
billiard==4.2.4
celery==5.6.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.1.0
frozenlist==1.8.0
gunicorn==23.0.0
hiredis==3.3.0
idna==3.10
iniconfig==2.3.0
joblib==1.5.2
kombu==5.6.2
lxml==6.0.2
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
